+++ cli/communicationToServer/communicationToServer.go
